update tests and types
